Add files via upload
polylines extraction.py reads a .xdf file (from e.g. Autodesk fusion 360) and generates a .csv file. As explained in chapter 6.1.1 in the thesis, this must be processed by hand in order to get the four required columns: t,x,y,z. Then copy the whole table, paste it into a text document and save in a folder called "meshes".
splinedata_nozzle.txt was generated with this python code and was used in the mesh generation of JetRIS_verysimp_nozzle,jl